changes from coda-oss

+++ externals/coda-oss/modules/c++/io/unittests/test_streams.cpp
@@ -77,10 +77,10 @@ TEST_CASE(testByteStream)
 
     stream.seek(2, io::Seekable::END);
     TEST_ASSERT_EQ(stream.tell(), 18);
-    TEST_ASSERT_EQ(stream.getSize(), 20);
+    TEST_ASSERT_EQ(stream.getSize(), static_cast<size_t>(20));
 
     stream.write("abcdef");
-    TEST_ASSERT_EQ(stream.getSize(), 24);
+    TEST_ASSERT_EQ(stream.getSize(), static_cast<size_t>(24));
 
     stream.clear();
     TEST_ASSERT_EQ(stream.available(), 0);
